requirements.txt
pandas==2.0.3
numpy==1.24.4
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.3.2
scikit-learn==1.3.2
huggingface-hub==0.20.2
pyarrow==14.0.2
torch==2.1.2
transformers==4.36.2
seaborn==0.13.0
datasets==2.16.1